remove action and add createGame on submit
